requirements.txt
langgraph
langchain-openai
python-dotenv
google-generativeai